data_generator/requirements.txt
psycopg2-binary==2.9.10
pandas==2.3.3
faker==18.10.0
pydantic-settings==2.10.1